Docs: switch to use new component data structure (#3054)
