+++ docs/docs-components/OverviewList.js
@@ -1,28 +1,13 @@
 // @flow strict
 import { type Node } from 'react';
-import { type Category, type ListItemType } from './data/components.js';
+import { type Category } from './data/components.js';
+import { type Platform, type PlatformData } from './data/types.js';
 import IllustrationCard from './IllustrationCard.js';
 import IllustrationSection from './IllustrationSection.js';
 
 const getIllustrationCardColor = (category: Category, hasDarkBackground?: boolean) => {
   const tealBackgrounds = ['Foundations'];
   const grayBackgrounds = ['Utilities', 'Building blocks'];
-  const greenBackgrounds = [
-    'Actions',
-    'Avatars',
-    'Controls',
-    'Data',
-    'Fields and forms',
-    'Help and guidance',
-    'Indicators',
-    'Loading',
-    'Messaging',
-    'Navigation',
-    'Overlays',
-    'Pins and imagery',
-    'Structure',
-    'Text',
-  ];
 
   if (hasDarkBackground) {
     return 'gray-roboflow-600';
@@ -36,27 +21,25 @@ const getIllustrationCardColor = (category: Category, hasDarkBackground?: boolea
     return 'gray-roboflow-100';
   }
 
-  if (greenBackgrounds.includes(category)) {
-    return 'green-matchacado-0';
-  }
-
   return 'green-matchacado-0';
 };
 
-export default function List({
-  array,
+type Props = {|
+  components: $ReadOnlyArray<PlatformData>,
+  headingLevel: 2 | 3,
+  platform: Platform,
+  title?: string,
+|};
+
+export default function OverviewList({
+  components,
   headingLevel,
   platform,
   title = '',
-}: {|
-  array: $ReadOnlyArray<ListItemType>,
-  headingLevel: 2 | 3,
-  platform: 'Web' | 'Android' | 'iOS',
-  title?: string,
-|}): Node {
+}: Props): Node {
   return (
     <IllustrationSection title={title} grid="auto-fill" min={312}>
-      {[...array]
+      {[...components]
         .sort((a, b) => {
           if (a.name < b.name) {
             return -1;
@@ -66,18 +49,17 @@ export default function List({
           }
           return 0;
         })
-        .map((element) => (
+        .map(({ name, visual: { svg, hasDarkBackground }, category, description, path }) => (
           <IllustrationCard
+            key={name}
+            // This is kind of a hacky assumption
+            // We should consider better ways to handle components with multiple categories
+            color={getIllustrationCardColor(category[0], hasDarkBackground)}
+            description={description}
             headingLevel={headingLevel}
-            key={element.name}
-            href={
-              element?.path ??
-              `/${platform.toLowerCase()}/${element.name.replace(/\s/g, '_').toLowerCase()}`
-            }
-            title={element.name}
-            description={element.description}
-            color={getIllustrationCardColor(element.category, element?.hasDarkBackground)}
-            image={element.svg}
+            href={path ?? `/${platform}/${name.replace(/\s/g, '_').toLowerCase()}`}
+            image={svg}
+            title={name}
           />
         ))}
     </IllustrationSection>
